requirements.txt
# XAI Project - Stability-Enhanced LIME for Russian Sentiment Analysis
# Baseline dependencies

# Core ML & NLP
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
scikit-learn>=1.3.0

# LIME and interpretability
lime>=0.2.0.1

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
